ref(browser): Streamline `showReportDialog` code (#16339)
Just a small refactor, noticed this while working on other browser sdk
stuff. Moving this into a dedicated file decouples this from the core
SDK functionality, which this IMHO is not (anymore). Also, we can
streamline the code slightly for efficiency.
